pyproject.toml
# my_all_in_one_dashboard/pyproject.toml

[tool.poetry]
name = "all-in-one-dashboard"
version = "1.0.0"
description = "An all-in-one dashboard with frontend and backend."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"